+++ hammer-core/src/test/java/cn/featherfly/hammer/tpl/mapper/freemarker/tpl_transverter2.sql
@@ -1,0 +1,7 @@
+select /*prop*/* from role
+/*<where*/ where
+/*??*/ name like :name
+/*??*/ or name like %:name1%
+/*??*/ or name like :name2%
+/*??*/ or name like %:name3
+/*>where*/