+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/tpl/freemarker/directive/WhereTemplateDirectiveModel.java
@@ -37,6 +37,8 @@ public class WhereTemplateDirectiveModel implements FreemarkerDirective, WhereDi
             if (condition.length() > 0) {
                 out.write(" where ");
                 out.write(condition);
+                //                out.write(condition.replaceAll("\n\n", "\n")); //TODO 后续加入配置来处理空白行的问题
+                out.write(" ");
             }
         }
     }